Adding ball detection to new main. Running Vision in different thread.

+++ RoboCup2016/GoalKeeper2016/Vision/DetectedObject.h
@@ -9,7 +9,7 @@
 class DetectedObject
 {
 public:
-	Point2i Center;
+	Point2f Center;
 	float Distance;
 
 	virtual bool IsDetected() = 0;
